backend/app/requirements.txt
fastapi==0.88.0
uvicorn==0.18.2
pydantic==1.10.2
pytest==7.2.2
httpx==0.24.0
passlib[bcrypt]==1.7.4
pymongo==4.3.3
python-dotenv==0.21.1
